[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -6,9 +6,9 @@ int main()
     ios::sync_with_stdio(0);
     cin.tie(0);
 
-    int a, b;
+    long long a, b;
     cin >> a >> b;
-    int c; 
+    long long c; 
     c = (a+b) * (a-b);
     cout << c;
 }
